Sovereignty Map Model changed to a more logical structure
